Formatting fixes.

+++ BundleFile/BundleCamera.h
@@ -6,20 +6,20 @@
 
 class BundleCamera
 {
-  float f, k1, k2;
-  Eigen::Matrix3f R;
-  Eigen::Vector3f t;
+  double f, k1, k2;
+  Eigen::Matrix3d R;
+  Eigen::Vector3d t;
   friend std::istream & operator>> (std::istream &, BundleCamera &);
   friend std::ostream & operator<< (std::ostream &, const BundleCamera &);
 public:
   BundleCamera():f(0),k1(0),k2(0){R.setZero(); t.setZero();}
-  inline float GetF() const { return f; }
-  inline float GetK1() const { return k1; }
-  inline float GetK2() const { return k2; }
-  inline const Eigen::Matrix3f & GetR() const { return R; }
-  inline const Eigen::Vector3f & GetT() const { return t; }
-  inline void SetR(const Eigen::Matrix3f & newR) { R = newR; }
-  inline void SetT(const Eigen::Vector3f & newT) { t = newT; }
+  inline double GetF() const { return f; }
+  inline double GetK1() const { return k1; }
+  inline double GetK2() const { return k2; }
+  inline const Eigen::Matrix3d & GetR() const { return R; }
+  inline const Eigen::Vector3d & GetT() const { return t; }
+  inline void SetR(const Eigen::Matrix3d & newR) { R = newR; }
+  inline void SetT(const Eigen::Vector3d & newT) { t = newT; }
   inline bool IsValid() const { return f > 0; }
 };
