+++ BundleFile/BundleView.h
@@ -6,16 +6,16 @@
 class BundleView
 {
   int camera, key;
-  float x, y;
+  double x, y;
   friend std::istream & operator>> (std::istream &, BundleView &);
   friend std::ostream & operator<< (std::ostream &, const BundleView &);
   public:
   BundleView () {}
-  BundleView (int camera, int key, float x, float y):camera(camera),key(key),x(x),y(y){}
+  BundleView (int camera, int key, double x, double y):camera(camera),key(key),x(x),y(y){}
   inline int GetCamera() const { return camera; }
   inline int GetKey() const { return key; }
-  inline float GetX() const { return x; }
-  inline float GetY() const { return y; }
+  inline double GetX() const { return x; }
+  inline double GetY() const { return y; }
 };
 
 #endif
